break downloader and documenter into separate files
