+++ src/index.js
@@ -1,0 +1,6 @@
+const documenter = require('./documenter');
+const downloader = require('./downloader');
+
+documenter.documenter = documenter; // for compoatibility
+documenter.downloader = downloader;
+module.exports = documenter;
